third_party/blink/web_tests/fast/css/getComputedStyle/getComputedStyle-list-style-shorthand.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>

description("Tests that the list-style shorthand is computed properly.")

var testContainer = document.createElement("div");
testContainer.contentEditable = true;
document.body.appendChild(testContainer);

testContainer.innerHTML = '<ul id="test"><li>Coffee</li><li>Tea</li><li>Coca Cola</li></ul>';

e = document.getElementById('test');
computedStyle = window.getComputedStyle(e, null);

e.style.listStyle = "circle outside url(dummy://test.png)";
shouldBeEqualToString("computedStyle.getPropertyValue('list-style')", 'outside url("dummy://test.png") circle');

e.style.listStyle = ""
e.style.listStyle = "a b c";
shouldBeEqualToString("computedStyle.getPropertyValue('list-style')", 'outside none disc');

e.style.listStyle = "circle outside";
shouldBeEqualToString("computedStyle.getPropertyValue('list-style')", 'outside none circle')

e.style.listStyle = "none";
shouldBeEqualToString("computedStyle.getPropertyValue('list-style')", 'outside none none')

e.style.listStyle = "";
e.style.listStyleType = "circle";
e.style.listStylePosition = "outside";
e.style.listStyleImage = "url(dummy://test.png)";
shouldBeEqualToString("computedStyle.getPropertyValue('list-style')", 'outside url("dummy://test.png") circle');

document.body.removeChild(testContainer);

</script>
</body>
</html>
